my_day/requirements.txt
asgiref==3.9.1
Django==5.2.6
djangorestframework==3.16.1
sqlparse==0.5.3
asgiref==3.9.1
Django==5.2.6
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
PyJWT==2.10.1
sqlparse==0.5.3
